refactor: simplify signup page structure and integrate progress indicator
